+++ lib/gl/vertexpipeline/Clipper.cpp
@@ -160,12 +160,12 @@ std::size_t Clipper::clipAgainstPlane(ClipList& listOut, const OutCode clipPlane
     // Start Clipping
     std::size_t i = 0;
 
-    for (int32_t vert = 0; vert < listSize; vert++)
+    for (int32_t vert = 0; vert < static_cast<int32_t>(listSize); vert++)
     {
         if (hasOutCode(listIn[vert].vertex, clipPlane))
         {
             // std::size_t vertMod = (vert - 1) % listSize;
-            const std::size_t vertPrev = (vert - 1) < 0 ? listSize - 1 : vert - 1;
+            const std::size_t vertPrev = (vert - 1) < 0 ? listSize - 1 : static_cast<std::size_t>(vert - 1);
             if (!hasOutCode(listIn[vertPrev].vertex, clipPlane))
             {
                 listOut[i] = lerp(clipPlane, listIn[vert], listIn[vertPrev]);
@@ -173,7 +173,7 @@ std::size_t Clipper::clipAgainstPlane(ClipList& listOut, const OutCode clipPlane
             }
 
             // vertMod = (vert + 1) % listSize;
-            const std::size_t vertNext = (vert + 1) >= listSize ? 0 : vert + 1;
+            const std::size_t vertNext = ((vert + 1) >= static_cast<int32_t>(listSize)) ? 0 : static_cast<std::size_t>(vert + 1);
             if (!hasOutCode(listIn[vertNext].vertex, clipPlane))
             {
                 listOut[i] = lerp(clipPlane, listIn[vert], listIn[vertNext]);
